+++ changelog.md
@@ -4,7 +4,7 @@ Liferay Plugin for IntelliJ
 All notable changes to this project will be documented in this file.
 
 ## 0.0.11
-
+- Feature: Implicit usage of some Language property keys like "javax.portlet.title." (no longer shown as "unused property")
 
 ## 0.0.10
 - Feature: Additional taglibs available in Freemarker templates for Liferay 7.1, 7.2 and 7.3